feat: concludes first crud (product and category)

+++ app/Controllers/Http/ProductsController.ts
@@ -3,16 +3,12 @@ import Product from 'App/Models/Product'
 import User from 'App/Models/User'
 import { schema } from '@ioc:Adonis/Core/Validator'
 import { PostValidator, UpdateValidator } from 'App/Validators/Products/Index'
+import Category from 'App/Models/Category'
 
 export default class ProductsController {
   // responsible for listing
   public async index({ params }: HttpContextContract) {
     const { slug, idCategory } = params
-    if (!slug) return Product.all()
-
-    console.log('%cslug: ', 'color: MidnightBlue; background: Aquamarine;', slug)
-    console.log('%cid: ', 'color: MidnightBlue; background: Aquamarine;', idCategory)
-
     const user = await User.findByOrFail('slug', slug)
 
     if (idCategory) {
@@ -23,35 +19,9 @@ export default class ProductsController {
       return products
     }
 
-    const products = await Product.query().where('idTenant', user!.id)
+    const productsAll = await Product.query().where('idTenant', user!.id)
 
-    return products
-  }
-
-  // responsible for store on data base
-  public async store({ request, auth }: HttpContextContract) {
-    const user = await auth.authenticate()
-    const data = await request.validate(PostValidator)
-    const imageData = await request.validate({
-      schema: schema.create({
-        image: schema.file.optional({
-          size: '5mb',
-          extnames: ['jpg', 'jpeg', 'png'],
-        }),
-      }),
-    })
-
-    if (imageData.image) {
-      try {
-        await imageData.image.moveToDisk('images', {}, 's3')
-        data.imageUrl = imageData.image.filePath
-      } catch (error) {
-        console.warn('%cerror: ', 'color: MidnightBlue; background: Aquamarine;', error)
-      }
-    }
-
-    const product = await Product.create({ idTenant: user.id, ...data })
-    return product
+    return productsAll
   }
 
   // responsible for showing something
@@ -63,17 +33,19 @@ export default class ProductsController {
     return product
   }
 
-  public async update({ request, params, auth }: HttpContextContract) {
-    // const user = await auth.authenticate()
+  // responsible for store on data base
+  public async store({ request, auth, response }: HttpContextContract) {
+    const tenant = await auth.authenticate()
+    const data = await request.validate(PostValidator)
 
-    const product = await Product.findOrFail(params.id)
-    const data = await request.validate(UpdateValidator)
+    const category = await Category.findOrFail(data.idCategory)
+    if (tenant.id !== category.idTenant) return response.unauthorized()
 
     const imageData = await request.validate({
       schema: schema.create({
         image: schema.file.optional({
           size: '5mb',
-          extnames: ['jpg', 'jpeg', 'png'],
+          extnames: ['jpg', 'jpeg', 'png', 'webp'],
         }),
       }),
     })
@@ -83,7 +55,36 @@ export default class ProductsController {
         await imageData.image.moveToDisk('images', {}, 's3')
         data.imageUrl = imageData.image.filePath
       } catch (error) {
-        console.warn('%cerror: ', 'color: MidnightBlue; background: Aquamarine;', error)
+        console.warn(error)
+      }
+    }
+
+    const product = await Product.create({ idTenant: tenant.id, ...data })
+    return product
+  }
+
+  public async update({ request, params, auth, response }: HttpContextContract) {
+    const tenant = await auth.authenticate()
+
+    const product = await Product.findOrFail(params.id)
+    if (tenant.id !== product.idTenant) return response.unauthorized()
+    const data = await request.validate(UpdateValidator)
+
+    const imageData = await request.validate({
+      schema: schema.create({
+        image: schema.file.optional({
+          size: '5mb',
+          extnames: ['jpg', 'jpeg', 'png', 'webp'],
+        }),
+      }),
+    })
+
+    if (imageData.image) {
+      try {
+        await imageData.image.moveToDisk('images', {}, 's3')
+        data.imageUrl = imageData.image.filePath
+      } catch (error) {
+        console.warn(error)
       }
     }
 
@@ -93,8 +94,12 @@ export default class ProductsController {
     return product
   }
 
-  public async destroy({ params }: HttpContextContract) {
-    const product = await Product.findOrFail(params.id)
+  public async destroy({ params, auth, response }: HttpContextContract) {
+    const { id } = params
+    const tenant = await auth.authenticate()
+    const product = await Product.findOrFail(id)
+
+    if (tenant.id !== product.idTenant) return response.unauthorized()
 
     await product.delete()
   }
